+++ lib/feature/about/presentation/presentation.dart
@@ -2,6 +2,7 @@ library about_presentation;
 
 export 'page/about_page.dart';
 export 'widgets/about_card.dart';
+export 'widgets/contact_cta.dart';
 export 'widgets/profile_card.dart';
 export 'widgets/skill_card.dart';
 export 'widgets/skill_icon.dart';