+++ lib/feature/feature.dart
@@ -4,5 +4,6 @@ export 'about/about.dart';
 export 'education/education.dart';
 export 'experience/experience.dart';
 export 'home/home.dart';
+export 'mail/mail.dart';
 export 'not_found/not_found.dart';
 export 'projects/projects.dart';