+++ lib/feature/mail/domain/service/email_service.dart
@@ -1,0 +1,8 @@
+abstract class EmailService {
+  Future<void> sendEmail({
+    required String title,
+    required String name,
+    required String email,
+    required String message,
+  });
+}