+++ lib/feature/mail/mail.dart
@@ -1,0 +1,10 @@
+library mail;
+
+export 'data/service/email_js_service_impl.dart';
+export 'domain/service/email_service.dart';
+export 'presentation/cubit/mail_cubit.dart';
+export 'presentation/cubit/mail_state.dart';
+export 'presentation/page/contact_page.dart';
+export 'presentation/widgets/contact_bullet_section.dart';
+export 'presentation/widgets/contact_me_form.dart';
+export 'presentation/widgets/contact_me_text_field.dart';